migrates to springboot
- migrates to Java 11
- runs Junit5
- upgrates spring-boot-starter-parent to 2.4.0
- includes properties for sonar
- includes spring-boot-starter-web
- includes jacoco-maven-plugin
- refactors path to xyz.wiraharas
- changes RequestMapping to GetMapping
