--- src/main/java/com/wiraharas/xyz/catatumbo/service/CuentaService.java
@@ -1,9 +1,0 @@
-package com.wiraharas.xyz.catatumbo.service;
-
-import com.wiraharas.xyz.catatumbo.entity.Cuenta;
-
-public interface CuentaService {
-
-    public Cuenta get(String rut);
-
-}
