--- src/main/java/com/wiraharas/xyz/catatumbo/service/TitularService.java
@@ -1,9 +1,0 @@
-package com.wiraharas.xyz.catatumbo.service;
-
-import com.wiraharas.xyz.catatumbo.entity.Titular;
-
-public interface TitularService {
-
-    public Titular get(String rut);
-
-}
