+++ src/main/java/xyz/wiraharas/controller/UsuarioController.java
@@ -1,24 +1,25 @@
-package com.wiraharas.xyz.catatumbo.controller;
+package xyz.wiraharas.controller;
 
-import com.wiraharas.xyz.catatumbo.entity.User;
-import com.wiraharas.xyz.catatumbo.service.UserService;
+import xyz.wiraharas.entity.User;
+import xyz.wiraharas.service.UserService;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
+@RequestMapping("/usuario")
 public class UsuarioController {
 
     @Autowired
     private UserService userService;
 
-    @RequestMapping(path = "/users", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-    public @ResponseBody
+    @GetMapping(path = "/",
+            produces = MediaType.APPLICATION_JSON_VALUE)
     List<User> listAllUsers() {
+
         return userService.listAll();
     }
