+++ src/main/java/xyz/wiraharas/service/CuentaService.java
@@ -1,0 +1,9 @@
+package xyz.wiraharas.service;
+
+import xyz.wiraharas.entity.Cuenta;
+
+public interface CuentaService {
+
+    public Cuenta get(String rut);
+
+}
