+++ src/main/java/xyz/wiraharas/service/TitularService.java
@@ -1,0 +1,9 @@
+package xyz.wiraharas.service;
+
+import xyz.wiraharas.entity.Titular;
+
+public interface TitularService {
+
+    public Titular get(String rut);
+
+}
