+++ src/main/resources/application.properties
@@ -1,0 +1,5 @@
+
+# Service configuration
+server.port=5155
+spring.application.name=wiraharas-xyz-service
+spring.profiles.active=testing
